refactor: add Timeout annotations and split ToolTest to ToolValidationTest
